Bug: Disk Array (#1)
The image.yaml, deployment.yaml, and data.yaml config files had the
deprecated style for disks. The disk section must be an array of disks
or else the gingervm commands will fail trying to parse the config file.

All the labels were removed because they are not being used.

Added an empty vars.json so the installation instructions could be
followed without any edits.

Added setup.sh so the directory matches what is in the installation
instructions.